+++ main.go
@@ -6,7 +6,7 @@ import (
 )
 
 //	@title			Device Management Service
-//	@version		0.4.127
+//	@version		0.4.128
 //	@description	A dashboard application for computing providers.
 //	@termsOfService	https://nunet.io/tos